Implemented SorghumState
Added more information to spline output for data generation.

Introduced SorghumState.
