+++ EvoEngine_Plugins/DigitalAgriculture/include/Sorghum.hpp
@@ -4,8 +4,10 @@ using namespace evo_engine;
 namespace digital_agriculture_plugin {
 class Sorghum final : public IPrivateComponent {
  public:
-  AssetRef sorghum_state_generator;
+  AssetRef sorghum_generator;
   AssetRef sorghum_growth_stages;
+
+  AssetRef sorghum_state;
   AssetRef sorghum_descriptor;
   void ClearGeometryEntities() const;
   void GenerateGeometryEntities(const SorghumMeshGeneratorSettings& sorghum_mesh_generator_settings);
