+++ EvoEngine_Plugins/DigitalAgriculture/include/SorghumModel/SorghumGrowthStages.hpp
@@ -1,17 +1,17 @@
 #pragma once
 #include "Curve.hpp"
 #include "Plot2D.hpp"
-#include "SorghumGrowthStage.hpp"
+#include "SorghumState.hpp"
 using namespace evo_engine;
 namespace digital_agriculture_plugin {
 class SorghumDescriptor;
 
 class SorghumGrowthStagePair {
-  void LeafStateHelper(SorghumLeafGrowthStage& left, SorghumLeafGrowthStage& right, float& a, int leaf_index) const;
+  void LeafStateHelper(SorghumLeafState& left, SorghumLeafState& right, float& a, int leaf_index) const;
 
  public:
-  SorghumGrowthStage left_stage = SorghumGrowthStage();
-  SorghumGrowthStage right_stage = SorghumGrowthStage();
+  SorghumState left_stage = SorghumState();
+  SorghumState right_stage = SorghumState();
   int state_mode = static_cast<int>(StateMode::Default);
   [[nodiscard]] int GetLeafSize(float a) const;
   [[nodiscard]] float GetStemLength(float a) const;
@@ -19,22 +19,22 @@ class SorghumGrowthStagePair {
   [[nodiscard]] glm::vec3 GetStemPoint(float a, float point) const;
   void ApplyPanicle(const std::shared_ptr<SorghumDescriptor>& target_state, float a) const;
   void ApplyStem(const std::shared_ptr<SorghumDescriptor>& target_state, float a) const;
-  void Apply(const std::shared_ptr<SorghumDescriptor>& target_state, float a) const;
-  void ApplyLeaves(const std::shared_ptr<SorghumDescriptor>& target_state, float a) const;
-  void ApplyLeaf(const std::shared_ptr<SorghumDescriptor>& target_state, float a, int leaf_index) const;
+  void Apply(const std::shared_ptr<SorghumDescriptor>& target_sorghum_descriptor, float a) const;
+  void ApplyLeaves(const std::shared_ptr<SorghumDescriptor>& target_descriptor, float a) const;
+  void ApplyLeaf(const std::shared_ptr<SorghumDescriptor>& target_descriptor, float a, int leaf_index) const;
 };
 
 class SorghumGrowthStages : public IAsset {
  public:
-  std::vector<std::pair<float, SorghumGrowthStage>> sorghum_growth_stages;
+  std::vector<std::pair<float, SorghumState>> sorghum_growth_stages;
   int state_mode = static_cast<int>(StateMode::Default);
   [[nodiscard]] bool ImportCsv(const std::filesystem::path& file_path);
   [[nodiscard]] float GetCurrentStartTime() const;
   [[nodiscard]] float GetCurrentEndTime() const;
-  void Add(float time, const SorghumGrowthStage& state);
+  void Add(float time, const SorghumState& state);
   void ResetTime(float previous_time, float new_time);
   void Remove(float time);
-  void Apply(const std::shared_ptr<SorghumDescriptor>& target_state, float time) const;
+  void Apply(const std::shared_ptr<SorghumDescriptor>& target_sorghum_descriptor, float time) const;
   bool OnInspect(const std::shared_ptr<EditorLayer>& editor_layer) override;
   void Serialize(YAML::Emitter& out) const override;
   void Deserialize(const YAML::Node& in) override;
