--- Resources/DigitalAgricultureProject/HeightFields/DigitalForestry.hf.evefilemeta
@@ -1,4 +1,0 @@
-asset_extension_: .hf
-asset_file_name_: DigitalForestry
-asset_type_name_: Binary
-asset_handle_: 1549083659852630678
