--- Resources/DigitalAgricultureProject/HeightFields/Flat.hf
@@ -1,22 +1,0 @@
-m_precisionLevel: 4
-m_noises2D:
-  m_minMax: [-1000, 1000]
-  m_noiseDescriptors:
-    - m_type: 2
-      m_frequency: 1
-      m_intensity: 1
-      m_multiplier: 0.0199999996
-      m_min: -10
-      m_max: 10
-      m_offset: 0
-      m_shift: [0, 0, 0]
-      m_ridgid: true
-    - m_type: 2
-      m_frequency: 8
-      m_intensity: 1
-      m_multiplier: 0.00999999978
-      m_min: -10
-      m_max: 10
-      m_offset: 0
-      m_shift: [0, 0, 0]
-      m_ridgid: false
