--- Resources/DigitalAgricultureProject/HeightFields/Flat.hf.evefilemeta
@@ -1,4 +1,0 @@
-asset_extension_: .hf
-asset_file_name_: Flat
-asset_type_name_: Binary
-asset_handle_: 12491391016094336752
