--- Resources/DigitalAgricultureProject/SoilLayers.evefoldermeta
@@ -1,2 +1,0 @@
-handle_: 8185060363503686886
-type_name: SoilLayers
