--- Resources/DigitalAgricultureProject/SoilLayers/Basic.nsld.evefilemeta
@@ -1,4 +1,0 @@
-asset_extension_: .nsld
-asset_file_name_: Basic
-asset_type_name_: Binary
-asset_handle_: 1161277276564892454
