--- Resources/DigitalAgricultureProject/SorghumDescriptorGenerator.evefoldermeta
@@ -1,2 +1,0 @@
-handle_: 762252338811437060
-type_name: SorghumDescriptorGenerator
