--- Resources/DigitalAgricultureProject/SorghumDescriptorGenerator/Random.sdg.evefilemeta
@@ -1,4 +1,0 @@
-asset_extension_: .sdg
-asset_file_name_: Random
-asset_type_name_: SorghumDescriptorGenerator
-asset_handle_: 15270584645888395836
