--- Resources/DigitalAgricultureProject/SorghumDescriptorGenerator/Season11.sdg.evefilemeta
@@ -1,4 +1,0 @@
-asset_extension_: .sdg
-asset_file_name_: Season11
-asset_type_name_: SorghumDescriptorGenerator
-asset_handle_: 7891237844097301122
