--- Resources/DigitalAgricultureProject/SorghumDescriptorGenerator/Season12.sdg.evefilemeta
@@ -1,4 +1,0 @@
-asset_extension_: .sdg
-asset_file_name_: Season12
-asset_type_name_: SorghumDescriptorGenerator
-asset_handle_: 14470078654346350886
