--- Resources/DigitalAgricultureProject/SorghumDescriptorGenerator/Skeleton.sdg.evefilemeta
@@ -1,4 +1,0 @@
-asset_extension_: .sdg
-asset_file_name_: Skeleton
-asset_type_name_: SorghumDescriptorGenerator
-asset_handle_: 10735921289881312219
