+++ Resources/DigitalAgricultureProject/SorghumGenerator/Random.sg.evefilemeta
@@ -1,0 +1,4 @@
+asset_extension_: .sg
+asset_file_name_: Random
+asset_type_name_: SorghumGenerator
+asset_handle_: 4826764104128816475
