+++ Resources/DigitalAgricultureProject/SorghumGenerator/Season12.sg.evefilemeta
@@ -1,0 +1,4 @@
+asset_extension_: .sg
+asset_file_name_: Season12
+asset_type_name_: SorghumGenerator
+asset_handle_: 1765328976826654380
