+++ Resources/DigitalAgricultureProject/SorghumGenerator/Skeleton.sg
