+++ Resources/DigitalAgricultureProject/SorghumGenerator/Skeleton.sg.evefilemeta
@@ -1,0 +1,4 @@
+asset_extension_: .sg
+asset_file_name_: Skeleton
+asset_type_name_: SorghumGenerator
+asset_handle_: 18060255872713625105
