+++ Resources/DigitalAgricultureProject/leafSurface.png.evefilemeta
@@ -1,4 +1,4 @@
 asset_extension_: .png
-asset_file_name_: LeafSurface
+asset_file_name_: leafSurface
 asset_type_name_: Texture2D
-asset_handle_: 9777172551836624152
+asset_handle_: 18194537419734825216
